Move core metrics registration into a new metriken-core crate
Part of the issue with how metriken is designed is that there can only
be a single copy of metriken within the whole dependency tree. This is
not great since it makes a major version update rather painful.

The solution I am adding here is to extract the core trait and linker
shenanigans into a separate metriken-core crate. The goal here is that
the metriken-core crate will never have a major version bump while the
metriken crate can have new versions when necessary.

To further that goal, I have made sure that the metriken-core API exposes
no types from any other crate.
