+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "histogram",
     "metriken",
     "metriken/derive",
+    "metriken/core",
     "ratelimit",
     "ringlog",
     "switchboard",
