+++ metriken/core/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "metriken-core"
+version = "0.1.0"
+edition = "2021"
+authors = [
+    "Brian Martin <brian@iop.systems>",
+    "Sean Lynch <sean@iop.systems>",
+]
+license = "Apache-2.0"
+homepage = "https://github.com/pelikan-io/rustcommon"
+repository = "https://github.com/pelikan-io/rustcommon"
+
+# Metrics are linked in metriken by using a linkme array.
+#
+# Having multiple versions of the same array results in a runtime panic so we
+# turn that into a compile-time error by using the links directive.
+links = "metriken-core"
+
+[dependencies]
+linkme = "0.3"
+parking_lot = "0.12"
+phf = "0.11"
+
