+++ metriken/core/build.rs
@@ -1,0 +1,2 @@
+// Empty build script so that we can use the links directive.
+fn main() {}
