+++ metriken/derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "metriken-derive"
-version = "0.4.0"
+version = "0.5.0"
 edition = "2021"
 authors = [
 	"Brian Martin <brian@pelikan.io>",
